# docs/swagger.yaml
basePath: /
definitions:
  account.Account:
    properties:
      accound_id:
        type: integer
      document_number:
        type: string
    type: object
  response.Error:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
  transaction.OperationType:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - Purchase
    - Installment
    - Withdrawal
    - Payment
  transaction.Transaction:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      id:
        type: integer
      operation_type_id:
        $ref: '#/definitions/transaction.OperationType'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing transactions
  title: Transaction APIs
  version: "1.0"
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Create a new account
      operationId: create-account
      parameters:
      - description: Account object to be created
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/account.Account'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Create a new account
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get an account
      tags:
      - accounts
  /transaction:
    post:
      consumes:
      - application/json
      description: Create a new transaction
      operationId: create-transaction
      parameters:
      - description: Transaction object to be created
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/transaction.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Create a new transaction
swagger: "2.0"